consiglieri menu

+++ Projeto2/src/main/java/com/mycompany/projeto2/Negocio.java
@@ -38,7 +38,7 @@ public class Negocio {
         return ValorBaseTributavel;
     }
 
-    public void setValorBaseTributavel(int NovoValorBase) {
+    public void setValorBaseTributavel(double NovoValorBase) {
         this.ValorBaseTributavel = NovoValorBase;
     }
